# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Shopping App - CRUD",
        "description": "My Shopping App Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "172.16.2.106:4000",
    "basePath": "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": ["Products"],
                "summary": "Get all products",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "query",
                        "required": false,
                        "description": "categoryId",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": ["Products"],
                "summary": "Create Product API",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "formData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "Category Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "productShortDescription",
                        "in": "formData",
                        "description": "Product Short Description",
                        "type": "string"
                    },
                    {
                        "name": "productDescription",
                        "in": "formData",
                        "description": "Product Description",
                        "type": "string"
                    },
                    {
                        "name": "productPrice",
                        "in": "formData",
                        "description": "Product Price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "productSalePrice",
                        "in": "formData",
                        "description": "Product Sale Price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "productSKU",
                        "in": "formData",
                        "description": "Product SKU",
                        "type": "string"
                    },
                    {
                        "name": "productType",
                        "in": "formData",
                        "description": "Product Type",
                        "type": "string"
                    },
                    {
                        "name": "productSatus",
                        "in": "formData",
                        "description": "Product Status",
                        "type": "string"
                    },
                    {
                        "name": "productImage",
                        "in": "formData",
                        "description": "Product Image",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse"
                        }
                    }
                }
            }
        },
        "/api/product/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "Product Id",
                    "type": "string"
                }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Product Details based on Product Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": ["Products"],
                "summary": "Update Product API",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "formData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "Category Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "productShortDescription",
                        "in": "formData",
                        "description": "Product Short Description",
                        "type": "string"
                    },
                    {
                        "name": "productDescription",
                        "in": "formData",
                        "description": "Product Description",
                        "type": "string"
                    },
                    {
                        "name": "productPrice",
                        "in": "formData",
                        "description": "Product Price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "productSalePrice",
                        "in": "formData",
                        "description": "Product Sale Price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "productSKU",
                        "in": "formData",
                        "description": "Product SKU",
                        "type": "string"
                    },
                    {
                        "name": "productType",
                        "in": "formData",
                        "description": "Product Type",
                        "type": "string"
                    },
                    {
                        "name": "productSatus",
                        "in": "formData",
                        "description": "Product Status",
                        "type": "string"
                    },
                    {
                        "name": "productImage",
                        "in": "formData",
                        "description": "Product Image",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Products"],
                "summary": "Delete Product API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Category": {
            "required": ["categoryName", "_id"],
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryDescription": {
                    "type": "string"
                },
                "categoryImage": {
                    "type": "string"
                }
            }
        },
...
        "ServiceResponse": {
            "properties": {
                "message": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Category"
                    }
                }
            }
        },
...
    }
}